moved functions from processor to main_utils
